refactor: authFetcher 관련 로직 해당 폴더로 이동
