+++ frontend/src/hooks/queries/profile/useMyPosts.ts
@@ -2,7 +2,7 @@ import { QueryKey, useQuery, UseQueryOptions } from 'react-query';
 
 import { AxiosError, AxiosResponse } from 'axios';
 
-import authFetcher from '@/apis';
+import authFetcher from '@/apis/authFetcher';
 import QUERY_KEYS from '@/constants/queries';
 
 type Size = number;
