+++ frontend/src/index.tsx
@@ -10,10 +10,6 @@ import { AuthContextProvider } from './context/Auth';
 import { PaginationContextProvider } from './context/Pagination';
 import { SnackBarContextProvider } from './context/Snackbar';
 
-import authFetcher from './apis';
-import { STORAGE_KEY } from './constants/localStorage';
-import { isExpired, parseJwt } from './utils/decodeJwt';
-
 import App from './App';
 import runJenniferFront from './jenniferFront';
 import GlobalStyle from './style/GlobalStyle';
@@ -46,20 +42,6 @@ axios.defaults.withCredentials = true;
 const queryClient = new QueryClient();
 
 const rootNode = document.getElementById('root') as Element;
-
-const refreshToken = localStorage.getItem(STORAGE_KEY.REFRESH_TOKEN);
-
-if (refreshToken && isExpired(parseJwt(refreshToken)!)) {
-  localStorage.removeItem(STORAGE_KEY.ACCESS_TOKEN);
-  localStorage.removeItem(STORAGE_KEY.REFRESH_TOKEN);
-}
-
-if (refreshToken && !isExpired(parseJwt(refreshToken)!)) {
-  authFetcher.defaults.headers.common['Refresh-Token'] = refreshToken;
-}
-
-const accessToken = localStorage.getItem(STORAGE_KEY.ACCESS_TOKEN);
-if (accessToken) authFetcher.defaults.headers.common['Authorization'] = accessToken;
 
 ReactDOM.createRoot(rootNode).render(
   <React.StrictMode>
